Telephony: Compilation fixes.

+++ src/com/android/mms/transaction/RetryScheduler.java
@@ -69,7 +69,7 @@ public class RetryScheduler implements Observer {
         TelephonyManager telephonyManager = (TelephonyManager)mContext
             .getSystemService(Context.TELEPHONY_SERVICE);
         if (telephonyManager != null) {
-            flag = telephonyManager.isDataPossibleForSubscription(subId,
+            flag = telephonyManager.isDataPossibleForSubscription((int)subId,
                     PhoneConstants.APN_TYPE_MMS);
         }
         Log.d(TAG, "isMmsDataConnectivityPossible = " + flag + "subId = " + subId);
